@samysweb 2022-02-05
Concerning the tests:
The command `coverage run` resulted in the following output:
```
42 failed, 1098 passed, 78 skipped, 28 xfailed, 273 warnings
```
However, this corresponds to the result when I'm running the test suite on the `develop` branch so I believe my commit does not break anything (new).
To be fair, the locally failing tests might be my fault though, because my BaB installation is currently broken...